backend/requirements.txt
Django==4.0
django-cors-headers==3.11.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
python-decouple==3.6
sqlparse==0.4.2
tzdata==2021.5
zipp==3.7.0
